Update Slider.js - Popup fixed
Fixed the position of the popup 8px right of the knob.
When the popup leaves the screen the popup balloon is mirrored and placed on 8px left of the knob.
the canvas of the popup is only redrawn when the direction of the balloon is changed.
The value controlColor is used for the background of the popup balloon. Else the color of the knob is used.